anomaly_detector/requirements.txt
grpcio-tools
numpy
scikit-learn
pandas
